rag_medical/requirements.txt
# ============================================================================
# DEPENDÊNCIAS DO PIPELINE RAG PARA DADOS MÉDICOS
# ============================================================================

# Pinecone - Vector Database
pinecone-client>=3.0.0

# LangChain - Framework RAG
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0

# Google Generative AI - Embeddings Gemini
google-generativeai>=0.8.0

# Gerenciamento de variáveis de ambiente
python-dotenv>=1.0.0

# Tokenização e processamento de texto
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=7.0.0

# Análise de dados (opcional, mas útil)
pandas>=2.0.0
numpy>=1.24.0

# Utilitários
tqdm>=4.65.0  # Barras de progresso
tenacity>=8.2.0  # Retry logic

